+++ apps/www/src/views/LineageView/LineageNodeView/index.tsx
@@ -1,6 +1,7 @@
 import { ImageListParameters, ImageWithEmbedded, Node } from "@phylopic/api-models"
-import { PaginationContainer } from "@phylopic/ui"
+import { Loader, PaginationContainer } from "@phylopic/ui"
 import { Query } from "@phylopic/utils"
+import clsx from "clsx"
 import Link from "next/link"
 import { FC, useMemo } from "react"
 import nodeHasOwnCladeImages from "~/models/nodeHasOwnCladeImages"
@@ -21,21 +22,25 @@ const LineageNodeView: FC<Props> = ({ value }) => {
     return (
         <section className={styles.main}>
             <PaginationContainer endpoint={`${process.env.NEXT_PUBLIC_API_URL}/images`} query={query}>
-                {items =>
+                {(items, _total, isLoading) =>
                     items.length ? (
-                        <div className={styles.images} key="images">
+                        <div className={styles.images}>
                             <ImageListView value={items as ImageWithEmbedded[]} />
+                        </div>
+                    ) : isLoading ? (
+                        <div className={clsx(styles.images, styles.loader)}>
+                            <Loader />
                         </div>
                     ) : null
                 }
             </PaginationContainer>
-            <header className={styles.header} key="header">
+            <header className={styles.header}>
                 {linked ? (
                     <Link href={`/nodes/${value.uuid}`}>
-                        <NomenView value={value.names[0]} short key="name" />
+                        <NomenView value={value.names[0]} short />
                     </Link>
                 ) : (
-                    <NomenView value={value.names[0]} short key="name" />
+                    <NomenView value={value.names[0]} short />
                 )}
             </header>
         </section>