+++ packages/ui/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.3.0] - 2023-03-26
+
+### Added
+
+-   Passing `isLoading` from `PaginationContainer`.
+
 ## [1.2.1] - 2023-03-24
 
 ### Added